+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## v0.1.9 (2022-09-06)
+  * Use only TLS 1.2 on OTP versions less than 25.
+
+## v0.1.8 (2022-07-14)
+  * Fix generated tailwind.config.js missing plugin reference
+
+## v0.1.7 (2022-07-13)
+  * Bump tailwindcss to v3.1.6
+  * Add Phoenix LiveView variants
+
+## v0.1.6 (2022-06-09)
+  * Bump tailwindcss to v3.1.0
+
 ## v0.1.5 (2022-01-18)
   * Prune app.js css import to remove required manual step on first install